+++ tensorflow/tools/api/golden/v1/tensorflow.data.-t-f-record-dataset.pbtxt
@@ -130,6 +130,10 @@ tf_class {
     argspec: "args=[\'self\', \'buffer_size\'], varargs=None, keywords=None, defaults=None"
   }
   member_method {
+    name: "random"
+    argspec: "args=[\'self\', \'seed\'], varargs=None, keywords=None, defaults=[\'None\'], "
+  }
+  member_method {
     name: "range"
     argspec: "args=[], varargs=args, keywords=kwargs, defaults=None"
   }
