+++ tensorflow/tools/api/golden/v2/tensorflow.data.-fixed-length-record-dataset.pbtxt
@@ -97,6 +97,10 @@ tf_class {
     argspec: "args=[\'self\', \'buffer_size\'], varargs=None, keywords=None, defaults=None"
   }
   member_method {
+    name: "random"
+    argspec: "args=[\'self\', \'seed\'], varargs=None, keywords=None, defaults=[\'None\'], "
+  }
+  member_method {
     name: "range"
     argspec: "args=[], varargs=args, keywords=kwargs, defaults=None"
   }
